Deprecate fabric module
